requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
bs4
langchain-text-splitters
langchain-openai
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain-chroma
streamlit
langchain_groq
langchain-core